cards v0.0.1 Cards

Provides methods for creating and handling a deck of cards

Functions

contains?(deck, card)

Checks if the List `deck` contains the Element `cards`

Examples

iex> deck = Cards.create_deck()
iex> Cards.contains?(deck, "Ace of Spades")
true
iex> Cards.contains?(deck, "Ace of")
false

create_deck()

Returns a List of Strings representing a deck of playing cards

create_hand(hand_size)

deal(deck, hand_size)

Takes a `deck`as an Argument and splits it into two lists `{hand, rest}`

hand wil have hand_sizes length.

Examples

    iex> deck = Cards.create_deck()
    iex> {hand, rest} = Cards.deal(deck, 1)
    iex> hand
    ["Ace of Spades"]

load(filename)

save(deck, filename)

shuffle(deck)

The List which will be given through the argument `deck` will be shuffled